Terrance/bso 88 refactor lopoglobals

+++ tests/utils/Defaults.sol
@@ -27,7 +27,7 @@ contract Defaults is Constants {
     uint256 public constant MINT_AMOUNT = 100_000e6;
     uint256 public constant COVER_AMOUNT = 10_000e6;
     uint256 public constant WITHDRAW_COVER_AMOUNT = 100e6;
-    uint256 public constant MIN_COVER_AMOUNT = 10e6;
+    uint104 public constant MIN_COVER_AMOUNT = 10e6;
     uint256 public constant REDEEM_AMOUNT = 1000e6;
     uint256 public constant PRINCIPAL = 100e6;
